fastapi/requirements.txt
# fastapi==0.63.0
# uvicorn==0.13.4
# gunicorn==20.1.0
# python-jose==3.2.0
# aiohttp==3.7.4
# sqlalchemy==1.3.16
# databases==0.3.2
# psycopg2==2.8.6
# asyncpg==0.22.0
# alembic==1.6.5
# passlib==1.7.4
# pydantic==1.8.2
# redis==3.5.3
# uvloop==0.15.2
# httptools==0.2.0
# cryptography==2.8
# python-multipart==0.0.5
# python-json-logger==2.0.1
# pytest==6.2.2
# pytest-asyncio==0.15.1
# pytest-cov==2.11.1
# flake8===3.8.4
# black==20.8b1
# isort==5.7.0
# gcloud-aio-storage==6.0.0
# google-auth==1.30.0
# openpyxl==3.0.7
# python-keycloak==0.25.0
# python-dotenv==0.17.0
# pytest-asyncio==0.15.1

fastapi==0.71.0
uvicorn==0.16.0
gunicorn==20.1.0
python-jose==3.3.0
aiohttp==3.8.1
sqlalchemy==1.4.5
alembic==1.7.5
asyncpg==0.25.0
aioresponses==0.7.2
nameparser==1.1.0
python-multipart==0.0.5
gevent==21.12.0
uvloop==0.16.0
httptools==0.3.0
python-dotenv==0.19.2
python-json-logger==2.0.2
pytest==6.2.5
requests==2.27.1
pytest-asyncio==0.16.0
pytest-cov==3.0.0
flake8==4.0.1
black==20.8b1
isort==5.10.1
gcloud-aio-storage==6.3.0
google-auth==2.3.3
openpyxl==3.0.9
Pillow==9.0.0
reportlab==3.6.5
aiofiles==0.8.0
python-keycloak==0.26.1
XlsxWriter==3.0.2
fpdf2==2.4.6
openpyxl==3.0.9
python-decouple==3.3
email-validator==1.1.2
PyJWT==1.7.1